Handle context cancels of pending connections
Previously if one side attempted to connect and the other side never
joined, the client would block waiting for rendezvous messages and
canceling would do nothing.

Now if the client is waiting for a message from the rendezvous server,
a context cancel will unblock the client and trigger an error.

Closes: #50 [via git-merge-pr]