test: case where server only accepts hashcash based permission
In order to support the test server, the "None" key in the json
message has been changed to accept a pointer. "omitempty" alone does
not work because it omits even if a "zero value" of that type is
assumed while unmarshalling, in this case, the zero value is
"struct{}{}", which is the value we accept as its normal value. We
really want to omit this field if server does not send it.